rename server.js to main.js
