+++ Procfile
@@ -1,2 +1,2 @@
-web: node src/server.js server
-scanner: node src/server.js scanner
+web: node src/main.js server
+scanner: node src/main.js scanner
